securing passphrase screens from taking screenshot
