+++ common/src/main/java/org/dash/wallet/common/InteractionAwareActivity.java
@@ -26,7 +26,7 @@ import android.view.WindowManager;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 
-public class InteractionAwareActivity extends AppCompatActivity {
+public class InteractionAwareActivity extends SecureActivity {
 
     public static final String FORCE_FINISH_ACTION = "InteractionAwareActivity.FORCE_FINISH_ACTION";
 
@@ -41,18 +41,6 @@ public class InteractionAwareActivity extends AppCompatActivity {
     public void onUserInteraction() {
         super.onUserInteraction();
         ((ResetAutoLogoutTimerHandler) getApplication()).resetAutoLogoutTimer();
-    }
-
-    @Override
-    protected void onPause() {
-        getWindow().addFlags(WindowManager.LayoutParams.FLAG_SECURE);
-        super.onPause();
-    }
-
-    @Override
-    protected void onResume() {
-        getWindow().clearFlags(WindowManager.LayoutParams.FLAG_SECURE);
-        super.onResume();
     }
 
     @Override
